leam-base.joda-time.mid-962.idx-170597.10.mutant

+++ joda-time/src/main/java/org/joda/time/MonthDay.java
@@ -888,7 +888,7 @@ public final class MonthDay
          */
         @Override
         public int get() {
-            return iBase.getValue(iFieldIndex);
+            return iBase.getValue((int) iFieldIndex);
         }
 
         //-----------------------------------------------------------------------
